백업파일 이동 및 최신내용 적용
